move StateStorageMode back
