+++ storage/src/main/java/tech/pegasys/teku/storage/server/StateStorageMode.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.storage.api;
+package tech.pegasys.teku.storage.server;
 
 public enum StateStorageMode {
