+++ teku/src/main/java/tech/pegasys/teku/cli/options/BeaconNodeDataOptions.java
@@ -20,8 +20,8 @@ import picocli.CommandLine.Option;
 import tech.pegasys.teku.beacon.sync.SyncConfig;
 import tech.pegasys.teku.config.TekuConfiguration;
 import tech.pegasys.teku.service.serviceutils.layout.DataConfig;
-import tech.pegasys.teku.storage.api.StateStorageMode;
 import tech.pegasys.teku.storage.server.DatabaseVersion;
+import tech.pegasys.teku.storage.server.StateStorageMode;
 import tech.pegasys.teku.storage.server.StorageConfiguration;
 
 public class BeaconNodeDataOptions extends ValidatorClientDataOptions {
