+++ teku/src/test/java/tech/pegasys/teku/cli/options/BeaconNodeDataOptionsTest.java
@@ -15,8 +15,8 @@ package tech.pegasys.teku.cli.options;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static tech.pegasys.teku.storage.api.StateStorageMode.ARCHIVE;
-import static tech.pegasys.teku.storage.api.StateStorageMode.PRUNE;
+import static tech.pegasys.teku.storage.server.StateStorageMode.ARCHIVE;
+import static tech.pegasys.teku.storage.server.StateStorageMode.PRUNE;
 
 import java.nio.file.Path;
 import org.junit.jupiter.api.Test;
